% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biomes-plot.R
\name{fun.poly.obj}
\alias{fun.poly.obj}
\title{fun.poly.obj}
\usage{
fun.poly.obj()
}
\value{
a SpatialPolygonsDataFrame of the biomes
}
\description{
Load biomes SpatialPolygons object
}
\details{
The Biomes are the Whittaker' biomes modified by Ricklefs (2008) in function
of mean annual temperature (MAT) and mean annual precipitation (MAP)
(MAT in degree Celsius and MAP in cm).
}
\author{
Kunstler
}
\references{
R.H. Whittaker. 1975. Communities and Ecosystems. 2d ed. Macmillan New York.

Ricklefs, R. E. (2008). The economy of nature. W. H. Freeman and Company. Chapter 5, Biological Communities, The biome concept.
}

