% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.transparent}
\alias{make.transparent}
\title{make.rgb.transparent}
\usage{
make.transparent(col, opacity = 0.5)
}
\arguments{
\item{col}{a vector of color}

\item{opacity}{percentage of opacity}
}
\value{
a vector of colors of length length(col)
}
\description{
make rgb color transparent
}
\details{
make rgb color transparent
}
\author{
Rich
}

