% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biomes-plot.R
\name{plot_biome}
\alias{plot_biome}
\title{plot.biome.map}
\usage{
plot_biome(poly = fun.poly.obj()$poly.DF,
  col.biomes = make.transparent(c("navajowhite3", "darkgoldenrod1", "sienna",
  "darkolivegreen4", "darkseagreen3", "forestgreen", "darkgreen", "olivedrab",
  "gray"), 0.9), add.legend = FALSE, add.number = TRUE,
  merge.deserts = FALSE, mar.vec = c(5.1, 3.5, 3, 2.1), cex.p = 1.2)
}
\arguments{
\item{poly}{biomes SpatialPolygons object}

\item{col.biomes}{vector of colors for the biomes (TODO add name of biomes)}

\item{add.legend}{TRUE or FALSE}

\item{add.number}{TRUE or FALSE}

\item{merge.deserts}{TRUE or FALSE}

\item{mar.vec}{vector of margine size}

\item{cex.p}{size of pointed passed to par}
}
\value{
plot
}
\description{
Plot the biomes
}
\details{
Takes the biomes SpatialPolygons object and plot it,
adding or not a legend
}
\author{
Kunstler
}

