% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_biome}
\alias{plot_biome}
\title{plot.biome.map}
\usage{
plot_biome(poly = fun.poly.obj()$poly.DF,
  col.biomes = make.transparent(c("navajowhite3", "darkgoldenrod1", "sienna",
  "darkolivegreen4", "darkseagreen3", "forestgreen", "darkgreen", "olivedrab",
  "gray"), 0.9), add.legend = TRUE)
}
\arguments{
\item{poly}{biomes SpatialPolygons object}

\item{col.biomes}{vector of colors for the biomes (TODO add name of biomes)}

\item{add.legend}{TRUE or FALSE}
}
\value{
plot
}
\description{
Plot the biomes
}
\details{
Takes the biomes SpatialPolygons object and plot it,
adding or not a legend
}
\author{
Kunstler
}

